Rewriting src/index.js - PART 1
Like it is being done with `c1-lighthouse`, code is being adapted to
comply with ESM and receiving and overall makeup job. Functionally
speaking, nothing changes, but looks are certainly improved.

Again `git` is working against me and recusing to accept all the
changes made to the file, so I'm dividing this is multiple commits.