Rewriting src/utils.js - Changing to jose from jsonwebtoken
Starting the serious part of the refactoring, this commit removes
`jsonwebtoken` from `c1-node` dependencies, changing to `jose` and
making it consistent with the rest of the programs.

When rewriting `generateKeyPair` and `createToken` I kept everything
perfectly compatible with what the rest of the codebase expects.

This commit also fix a somewhat serious bug: on its first run,
`getKeyPair` was not creating `shared` directory, because on the
abscence of `key` directory, the key pair was generated and then
returned immediately, skipping the `shared` dir bootstrap.