src directory flattening, segmentation comments, ESM by default
Excessive nesting brings messiness, so I'm flattening everything
inside `src` dir. When the time comes, if necessary, we group related
files together in their own subdirs.

I also modified every file with the minimum necessary changes to make
this project use ESM by default. This may be a polemic change, but if
the team disagrees it's easy enough to revert.