Making editors happy and consistent
Adding `jsconfig.json` so that every editor becomes well behaved and
understand that this is, in fact, a project and all the files inside
are related to one another. The file was copied from executing `bun
init` and making some minor changes.

Also adding VS Code workspace settings, `.vscode` dir, so that
everyone that uses it avoids messing up the code with unnecessary
whitespaces. Diffs will thank us.