peers.js rewrite DONE
Okay, here we have a complicated case. This file is dealing with the
"meat" of the project and so needs more study and tests to understand
how to improve and add new features.

Said that, this commit limits itself to applying code makeup, making
everything follow the current style line. Apart from this, I also
change `createPeer` to simply `create`, because it looks nicer when
using the module. Of course this can be reverted easily later.